Update README.md
Added initial github repository. Won't have much time to work on actual code until the weekend, but it is good to start the repository going and to think a bit about the desired features, architecture and tech stack.

At this point in time, it seems clear that an API Gateway may be built, as well as a user and store microservices. VueJS for front-end, since that's what I saw is being used by Jumbo as of late (medium post). Google map API will probably be used to calculate the distance.